FIX/CHORE: Type fixes, route security, test suite updates
reconcileSubscription amount type int→float (payments).
Add committee middleware to admin/website routes.
Add 5 test files (Flux B pricing, enrollment, notifications, lifecycle)
and update existing tests for enrolled-status pivot and price format.

+++ resources/views/pages/club-admin/treasury/⚡payments/payments.php
@@ -501,7 +501,7 @@ new class extends Component
             ->values();
     }
 
-    private function reconcileSubscription(Subscription $subscription, int $amount): void
+    private function reconcileSubscription(Subscription $subscription, float $amount): void
     {
         $subscription->update(['amount_paid' => $amount]);
